Sparkline:Migrated engyme tests to react tests
