+++ packages/charts/react-charts/library/src/components/Sparkline/SparklineChartUT.test.tsx
@@ -1,12 +1,56 @@
 import { getByClass, testWithWait } from '../../utilities/TestUtility.test';
 import { Sparkline } from './Sparkline';
-
-import { emptySparklinePoints, sparkline1Points } from './Sparkline.test';
+import { ChartProps } from './index';
+import { emptySparklinePoints } from './Sparkline.test';
 import { getByRole, queryAllByAttribute, render } from '@testing-library/react';
 import * as React from 'react';
 
 const env = require('../../../config/tests.js');
 const runTest = env === 'TEST' ? describe : describe;
+
+const sparkline1Points: ChartProps = {
+  chartTitle: '10.21',
+  lineChartData: [
+    {
+      legend: '19.64',
+      color: '#00AA00',
+      data: [
+        {
+          x: 1,
+          y: 58.13,
+        },
+        {
+          x: 2,
+          y: 140.98,
+        },
+        {
+          x: 3,
+          y: 20,
+        },
+        {
+          x: 4,
+          y: 89.7,
+        },
+        {
+          x: 5,
+          y: 99,
+        },
+        {
+          x: 6,
+          y: 13.28,
+        },
+        {
+          x: 7,
+          y: 31.32,
+        },
+        {
+          x: 8,
+          y: 10.21,
+        },
+      ],
+    },
+  ],
+};
 
 const sl1 = [
   {
